fix: don't capture spawn and fork output unless encoding or maxBuffer is given
fix #1

BREAKING CHANGE: If you want `stdout` or `stderr` in the result of a `spawn`ed or `fork`ed child
process, you must now pass the `maxBuffer` or `encoding` option.  (`maxBuffer` defaults to
`200 * 1024` if only `encoding` is give).  In previous versions all output
was always captured if piped to the parent, potentially exhausting available memory and crashing
the parent process.